Real-time formatting of yen input (#31)
* Real-time formatting of yen input

Uses the react-number-format library to format the input that takes yen amounts in real time, improving the usability and eliminating the need to use various features to work around this, such as the input adornment for the yen symbol and JPY, as well as the formatted helper text which was redundant info other than formatting.

* Use inputmode=numeric for better mobile experience

Against the advice of MUI, use input type numeric for a better mobile experience.
The downsides mentioned are largely mitigated with the react-number-format library.

* Test for numeric inputmode