Take employment/other income input
Replaces the net income range input with two input fields. One for gross employment income and one for total other net income. People are likely to not calculate net income correctly, so we seek input like this to better ensure correct results. We calculate the net employment income and add it to their other net income.

This also shows the result of what deductions the dependent makes the taxpayer eligible for and in what amounts for income tax purposes and residence tax purposes.