Support dependents-related deductions (#112)
Adds support for inputting necessary information about dependents to determine the taxpayer's eligibility for dependents-related deductions. Eligible deductions are included in the tax calculation and summarized for the user in both the dependents management screen and the breakdown results for taxes. A tooltip for the dependent-related deductions gives official sources so users can check the rules themselves.