Manual social insurance entry feature (#143)
While for many people the current calculations of social insurance premiums may be accurate, there are a number of situations that would make them inaccurate. To list some:

- Employees who do not have a fixed annual salary throughout the year
- Employees who receive bonuses
- Employees who receive RSU/ESPP compensation issued by a foreign company
- People on NHI who have a different income this year vs last year (though we can solve this by asking for last year's income)
- People on NHI living in a municipality not covered by options available in the dropdown
- People who have components of the [social insurance premium deduction](https://www.nta.go.jp/taxes/shiraberu/taxanswer/shotoku/1130.htm) we don't support (or don't pay employment insurance as an employee)

These users will know their total annual social insurance premiums either from 1) their annual withholding slip (源泉徴収票) if they're an employee or 2) from the number they report on their tax return. By allowing them to manually enter it in kei3, the other calculations can be kept accurate for that user without trying to add input options to support all of the above cases and more. We can continue to support more cases in the future and give users less reasons to need to use this manual override, but it's still better to keep this as an option to support more users now.

The chart data is effectively disabled when manually entering social insurance amount because we have no way to scale the manually entered amount to other incomes.