Age range selection using segemented buttons in input form
The toggle switch before was not following material design best practice, which is that it be used for a binary choice of something that is off or on. In some sense, the choice is that in that it is off/on for the extra premiums people aged 40-64 pay, but it is not exactly presented that way. The way it is presented, the best practice is to use segmented buttons to illustrate one and only one of the options is chosen.